Resolve some bugs.

+++ streamis-jobmanager/streamis-job-manager/streamis-job-manager-service/src/main/scala/com/webank/wedatasphere/streamis/jobmanager/manager/transform/builder/StreamisFlinkSqlTransformJobBuilder.scala
@@ -22,7 +22,7 @@ class StreamisFlinkSqlTransformJobBuilder extends StreamisTransformJobBuilder {
   @Autowired private var streamJobMapper: StreamJobMapper = _
   @Autowired private var configurationService: ConfigurationService = _
 
-  override def canBuild(streamJob: StreamJob): Boolean = JOBMANAGER_FLINK_SQL.getValue == streamJob.getJobType
+  override def canBuild(streamJob: StreamJob): Boolean = JOBMANAGER_FLINK_SQL.getValue == streamJob.getJobType.intValue()
 
   override def build(streamJob: StreamJob): StreamisTransformJob = {
     val transformJob = new StreamisSqlTransformJob
